Re-organise the code and split files out
