+++ gfort2py/fParameters.py
@@ -1,0 +1,21 @@
+class fParam:
+    def __init__(self, obj):
+        self.obj = obj
+
+    @property
+    def value(self):
+        return self.obj.value()
+
+    @value.setter
+    def value(self, value):
+        raise AttributeError("Parameters can't be altered")
+
+    def __repr__(self):
+        return repr(self.value)
+
+    def __str__(self):
+        return str(self.value)
+
+    @property
+    def module(self):
+        return self._value.module
